Rename signature/signature_utilities.py to rules.py
